fix: 중고거래-판매or예약 중인 내역 조회 구현 완료

+++ itemrier_boot/src/main/java/com/hotSix/itemrier_boot/dto/user/UserDto.java
@@ -9,7 +9,7 @@ import lombok.*;
 @NoArgsConstructor(force = true)
 @Data
 public class UserDto {
-    private Long userId;
+    private int userId;
 
 	private String email;
