+++ itemrier_boot/src/main/java/com/hotSix/itemrier_boot/repository/user/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import com.hotSix.itemrier_boot.domain.user.UserEntity;
 
 @Repository
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
+public interface UserRepository extends JpaRepository<UserEntity, Integer> {
 
-	UserEntity findByUserId(Long userId);
+	UserEntity findByUserId(int userId);
 }
